+++ business_requirement/__openerp__.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 # © 2016 Elico Corp (https://www.elico-corp.com).
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
 {
     'name': 'Business Requirement',
     'category': 'Business Requirements Management',
@@ -8,7 +9,7 @@
                 for your customers',
     'version': '9.0.1.0.0',
     'website': 'www.elico-corp.com',
-    "author": "Elico Corp, Odoo Community Association (OCA)",
+    'author': 'Elico Corp, Odoo Community Association (OCA)',
     'depends': [
         'project',
     ],
@@ -16,8 +17,10 @@
         'data/business_data.xml',
         'security/security.xml',
         'security/ir.model.access.csv',
-        'views/business_view.xml',
+        'views/business_requirement_category_view.xml',
+        'views/business_requirement_view.xml',
         'views/project.xml',
+        'menu_items.xml',
     ],
     'image': [
         'static/description/icon.png',
