+++ business_requirement/data/business_data.xml
@@ -10,7 +10,7 @@
             <field name="company_id" eval="False"/>
         </record>
 
-        <record model="ir.module.category" id="module_category_business_requirement">
+        <record id="module_category_business_requirement" model="ir.module.category">
             <field name="name">Business Requirement</field>
             <field name="description"></field>
             <field name="sequence">9</field>
