+++ business_requirement/models/__init__.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
 # © 2016 Elico Corp (https://www.elico-corp.com).
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
-from . import business
+
+from . import business_requirement
+from . import business_requirement_category
 from . import project
