+++ business_requirement/models/business_requirement_category.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+# © 2016 Elico Corp (https://www.elico-corp.com).
+# License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
+
+from openerp import api, fields, models
+
+
+class BusinessRequirementCategory(models.Model):
+    _name = "business.requirement.category"
+    _description = "Categories"
+
+    name = fields.Char(required=True)
+    parent_id = fields.Many2one(
+        comodel_name='business.requirement.category',
+        string='Parent Category',
+        ondelete='restrict'
+    )
