+++ business_requirement/security/ir.model.access.csv
@@ -1,6 +1,5 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
 access_business_requirement_user,business.requirement.user,model_business_requirement,business_requirement.group_business_requirement_user,1,1,1,1
 access_business_requirement_manager,business.requirement.manager,model_business_requirement,business_requirement.group_business_requirement_manager,1,1,1,1
-
 access_business_requirement_category_user,business.requirement.category.user,model_business_requirement_category,business_requirement.group_business_requirement_user,1,1,1,1
 access_business_requirement_category_manager,business.requirement.category.manager,model_business_requirement_category,business_requirement.group_business_requirement_manager,1,1,1,1
