+++ business_requirement/views/business_requirement_category_view.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="view_business_requirement_category_form" model="ir.ui.view">
+        <field name="name">business.requirement.category.form</field>
+        <field name="model">business.requirement.category</field>
+        <field name="priority">20</field>
+        <field name="arch" type="xml">
+            <form string="Business Requirement">
+                <sheet>
+                    <group col="4">
+                         <field name="name"/>
+                         <field name="parent_id"/>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+
+    <record id="view_business_requirement_category_tree" model="ir.ui.view">
+        <field name="name">business.requirement.category.form</field>
+        <field name="model">business.requirement.category</field>
+        <field name="priority">20</field>
+        <field name="arch" type="xml">
+            <tree string="Business Requirement">
+                 <field name="name"/>
+                 <field name="parent_id"/>
+            </tree>
+        </field>
+    </record>
+
+    <record id="action_business_requirement_category_tree" model="ir.actions.act_window">
+        <field name="name">Categories</field>
+        <field name="res_model">business.requirement.category</field>
+        <field name="view_mode">tree,form,graph</field>
+        <field name="view_type">form</field>
+        <field name="context">{}</field>
+    </record>
+
+</odoo>
