+++ business_requirement/views/business_requirement_view.xml
@@ -123,34 +123,6 @@
         </field>
     </record>
 
-    <record id="view_business_requirement_category_form" model="ir.ui.view">
-        <field name="name">business.requirement.category.form</field>
-        <field name="model">business.requirement.category</field>
-        <field name="priority">20</field>
-        <field name="arch" type="xml">
-            <form string="Business Requirement">
-                <sheet>
-                    <group col="4">
-                         <field name="name"/>
-                         <field name="parent_id"/>
-                    </group>
-                </sheet>
-            </form>
-        </field>
-    </record>
-
-    <record id="view_business_requirement_category_tree" model="ir.ui.view">
-        <field name="name">business.requirement.category.form</field>
-        <field name="model">business.requirement.category</field>
-        <field name="priority">20</field>
-        <field name="arch" type="xml">
-            <tree string="Business Requirement">
-                 <field name="name"/>
-                 <field name="parent_id"/>
-            </tree>
-        </field>
-    </record>
-
     <record id="act_view_project_br_all" model="ir.actions.act_window">
         <field name="name">Business Requirement</field>
         <field name="res_model">business.requirement</field>
@@ -188,38 +160,5 @@
         <field name="view_type">form</field>
         <field name="context">{}</field>
     </record>
-
-    <record id="action_business_requirement_category_tree" model="ir.actions.act_window">
-        <field name="name">Categories</field>
-        <field name="res_model">business.requirement.category</field>
-        <field name="view_mode">tree,form,graph</field>
-        <field name="view_type">form</field>
-        <field name="context">{}</field>
-    </record>
-
-    <!-- Top menu item -->
-    <menuitem id="menu_business_root" name="Business Requirement" 
-              groups="business_requirement.group_business_requirement_user" sequence="30"/>
-
-    <menuitem id="menu_business_requirement" name="Business Requirement"
-              groups="business_requirement.group_business_requirement_user"
-              parent="menu_business_root" sequence="8"/>
-
-    <menuitem id="menu_business_configuration" name="Configuration"
-              groups="business_requirement.group_business_requirement_user"
-              parent="menu_business_root" sequence="8"/>
-
-    <menuitem id="menu_business_requirement1"
-              groups="business_requirement.group_business_requirement_user"
-              action="action_business_requirement_tree" 
-              parent="menu_business_requirement" sequence="10"/>
-
-    <menuitem id="menu_business_requirement_category"
-              groups="business_requirement.group_business_requirement_user"
-              action="action_business_requirement_category_tree" 
-              parent="menu_business_configuration" sequence="10"/>
-
-    <menuitem action="product.product_template_action" id="menu_product_template_action"
-              parent="menu_business_configuration" sequence="11" />
 
 </odoo>
