+++ business_requirement/views/project.xml
@@ -20,17 +20,6 @@
         </field>
     </record>
 
-    <record id="open_view_subproject_all" model="ir.actions.act_window">
-        <field name="name">Projects</field>
-        <field name="res_model">project.project</field>
-        <field name="view_type">form</field>
-        <field name="view_mode">kanban,tree,form,gantt</field>
-        <field name="view_id" ref="view_project_kanban"/>
-        <field name="search_view_id" ref="project.view_project_project_filter"/>
-        <field name="context">{'search_default_parent_id': context.get('analytic_account_id', False)}
-        </field>
-    </record>
-
     <record id="edit_project" model="ir.ui.view">
         <field name="name">project.project.form</field>
         <field name="model">project.project</field>
@@ -47,6 +36,17 @@
                            widget="statinfo"/>
                 </button>
             </xpath>
+        </field>
+    </record>
+
+    <record id="open_view_subproject_all" model="ir.actions.act_window">
+        <field name="name">Projects</field>
+        <field name="res_model">project.project</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">kanban,tree,form,gantt</field>
+        <field name="view_id" ref="view_project_kanban"/>
+        <field name="search_view_id" ref="project.view_project_project_filter"/>
+        <field name="context">{'search_default_parent_id': context.get('analytic_account_id', False)}
         </field>
     </record>
